refactored discover section; separated out components
